+++ package.json
@@ -1,35 +1,35 @@
 {
-    "name": "node-red-contrib-homekit-enhanced",
-    "version": "1.0.0",
-    "description": "Node-RED nodes to simulate Apple HomeKit devices. Based on node-red-contrib-homekit-bridged.",
-    "main": "homekit.js",
-    "scripts": {
-        "test": "test"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/Shaquu/node-red-contrib-homekit-enhanced.git"
-    },
-    "keywords": [
-        "homekit",
-        "hap-nodejs",
-        "node-red",
-        "iot",
-        "homebridge",
-        "home"
-    ],
-    "node-red": {
-        "nodes": {
-            "homekit": "homekit.js"
-        }
-    },
-    "author": "Marius Schmeding, Oliver Rahner, Shaquu",
-    "license": "Apache-2.0",
-    "bugs": {
-        "url": "https://github.com/Shaquu/node-red-contrib-homekit-enhanced/issues"
-    },
-    "homepage": "https://github.com/Shaquu/node-red-contrib-homekit-enhanced#readme",
-    "dependencies": {
-        "hap-nodejs": "^0.4.48"
+  "name": "node-red-contrib-homekit-enhanced",
+  "version": "1.0.1",
+  "description": "Node-RED nodes to simulate Apple HomeKit devices. Based on node-red-contrib-homekit-bridged.",
+  "main": "homekit.js",
+  "scripts": {
+    "test": "test"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Shaquu/node-red-contrib-homekit-enhanced.git"
+  },
+  "keywords": [
+    "homekit",
+    "hap-nodejs",
+    "node-red",
+    "iot",
+    "homebridge",
+    "home"
+  ],
+  "node-red": {
+    "nodes": {
+      "homekit": "homekit.js"
     }
+  },
+  "author": "Marius Schmeding, Oliver Rahner, Shaquu",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/Shaquu/node-red-contrib-homekit-enhanced/issues"
+  },
+  "homepage": "https://github.com/Shaquu/node-red-contrib-homekit-enhanced#readme",
+  "dependencies": {
+    "hap-nodejs": "^0.4.48"
+  }
 }
